# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Ingestion & Privacy Intelligence API",
    "version": "1.0.0",
    "description": "Proteccio Discover backend covering Week 1 ingestion, Week 2 discovery/classification/mapping/profiling, and Week 3 governance/compliance intelligence: privacy risk, compliance controls, remediation workflows, dashboards, reporting, RBAC, alerting, and global search.\n\n**Auth:** When `JWT_SECRET` is set, use `POST /api/auth/login` and send `Authorization: Bearer <JWT>`. When only `API_KEY` is set, send `Authorization: Bearer <API_KEY>` or `X-API-Key: <API_KEY>` on `/api/*` except `/health` and documentation routes.\n\n**Docs UI:** `GET /docs` (Swagger UI). **Raw spec:** `GET /openapi.json`."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyHeader": []
    }
  ],
  "paths": {
...
    "/api/profiling/profile": {
      "post": {
        "tags": [
          "Profiling"
        ],
        "summary": "Build profiling (+ optional risk + catalog persist)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "profile, risk, optional catalog",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/register": {
      "post": {
        "tags": [
          "Profiling"
        ],
        "summary": "Register snapshot in catalog",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT from POST /api/auth/login when JWT_SECRET is set; otherwise API_KEY value."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same value as API_KEY when set."
      }
    },
    "schemas": {
...
      "ProfilingRequest": {
        "type": "object",
        "required": [
          "discovery"
        ],
        "properties": {
          "discovery": {
            "type": "object"
          },
          "classification": {
            "type": "object"
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "persist": {
            "type": "boolean"
          },
          "profilingOptions": {
            "type": "object"
          },
          "exposureHints": {
            "$ref": "#/components/schemas/ExposureHints"
          }
        }
      },
      "ComplianceControlHints": {
        "type": "object",
        "description": "Governance control attestations used by the compliance intelligence engine.",
        "properties": {
          "retentionPolicyIndicated": {
            "type": "boolean"
          },
          "consentManagementIndicated": {
            "type": "boolean"
          },
          "privacyNoticeIndicated": {
            "type": "boolean"
          },
          "lawfulBasisDocumented": {
            "type": "boolean"
          },
          "accessControlsIndicated": {
            "type": "boolean"
          },
          "breachNotificationProcessIndicated": {
            "type": "boolean"
          },
          "dataPrincipalRightsProcessIndicated": {
            "type": "boolean"
          },
          "baaInPlace": {
            "type": "boolean"
          },
          "phiAuditLoggingIndicated": {
            "type": "boolean"
          },
          "optOutMechanismIndicated": {
            "type": "boolean"
          },
          "ismsRiskAssessmentIndicated": {
            "type": "boolean"
          },
          "purposeLimitationDocumented": {
            "type": "boolean"
          },
          "crossBorderSafeguardsIndicated": {
            "type": "boolean"
          },
          "consumerDisclosureIndicated": {
            "type": "boolean"
          },
          "ismsDocumented": {
            "type": "boolean"
          }
        }
      },
      "ExposureHints": {
        "type": "object",
        "properties": {
          "hasApiExposureFlow": {
            "type": "boolean"
          },
          "hasReplicationOrBackupFlow": {
            "type": "boolean"
          },
          "isPubliclyExposed": {
            "type": "boolean"
          },
          "encryptionIndicated": {
            "type": "boolean"
          },
          "crossDatasetDuplicateGroupCount": {
            "type": "integer",
            "minimum": 0
          },
          "unmappedDataset": {
            "type": "boolean"
          },
          "noLineageFlows": {
            "type": "boolean"
          },
          "daysSinceLastActivity": {
            "type": "integer",
            "minimum": 0
          },
          "complianceControls": {
            "$ref": "#/components/schemas/ComplianceControlHints"
          }
        }
      },
...
    },
...
  }
}